requirements.txt
fastapi[standard]
uvicorn
sqlalchemy[asyncio]
aiosqlite
passlib
bcrypt==4.0.1
python-jose
dotenv